Fix bluetooth settings issues.
NimbleController.cpp was running bleController.Disconnect() each
time NimbleController::EnableRadio() was called by SystemTask.cpp's
case matching Messages::BleRadioEnableToggle. This meant that each time
the bluetooth settings page was opened, the Messages::BleRadioEnableToggle
case was matched and the setting variable bleRadioEnabled was set to false.
This caused the conditionals for the checkboxList class call in SettingBluetooth.cpp
to not allow toggling the bluetooth on and off, and caused the bluetooth status icon
to dissappear in StatusIcons.cpp